Part 6, 8/8. Sending emails doesn't work!

+++ routes/courses.js
@@ -3,26 +3,40 @@ const Course = require('../models/course')
 const auth = require('../middleware/auth')
 const router = Router()
 
-router.get('/', async (req, res) => {
-  const courses = await Course.find()
-    .populate('userId', 'email name')
-    .select('price title img')
+// является ли владельцем курса авторизованный пользователь
+function isOwner(course, req) {
+  return course.userId.toString() === req.user._id.toString()
+}
 
-  res.render('courses', {
-    title: 'Курсы',
-    isCourses: true,
-    courses
-  })
+router.get('/', async (req, res) => {
+  try {
+    const courses = await Course.find()
+      .populate('userId', 'email name')
+      .select('price title img')
+  
+    res.render('courses', {
+      title: 'Курсы',
+      isCourses: true,
+      userId: req.user ? req.user._id.toString() : null,
+      courses
+    })    
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 router.get('/:id', async (req, res) => {
-  const course = await Course.findById(req.params.id)
-
-  res.render('course', {
-    layout: 'empty',
-    title: `Курс ${course.title}`,
-    course
-  })
+  try {
+    const course = await Course.findById(req.params.id)
+  
+    res.render('course', {
+      layout: 'empty',
+      title: `Курс ${course.title}`,
+      course
+    })    
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 router.get('/:id/edit', auth, async (req, res) => {
@@ -30,24 +44,42 @@ router.get('/:id/edit', auth, async (req, res) => {
     return res.redirect('/')
   }
 
-  const course = await Course.findById(req.params.id)
+  try {
+    const course = await Course.findById(req.params.id)
 
-  res.render('course-edit', {
-    title: `Редактировать ${course.title}`,
-    course
-  })
+    if (!isOwner(course, req)) {
+      return res.redirect('/courses')
+    }
+
+    res.render('course-edit', {
+      title: `Редактировать ${course.title}`,
+      course
+    })
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 router.post('/edit', auth, async (req, res) => {
-  const {id} = req.body
-  delete req.body.id // в Mongo DB id не передаем
-  await Course.findByIdAndUpdate(id, req.body)
-  res.redirect('/courses')
+  try {
+    const {id} = req.body
+    delete req.body.id // в Mongo DB id не передаем
+    const course = await Course.findById(id)
+    if (!isOwner(course, req)) {
+      return res.redirect('/courses')
+    }
+    // await Course.findByIdAndUpdate(id, req.body)
+    Object.assign(course, req.body)
+    await course.save()
+    res.redirect('/courses')
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 router.post('/remove', auth, async (req, res) => {
   try {
-    await Course.deleteOne({ _id: req.body.id })
+    await Course.deleteOne({ _id: req.body.id, userId: req.user._id })
     res.redirect('/courses')
   } catch (error) {
     console.log(error)
